HAWQ-596. change gucs hawq_rm_tolerate_nseg_limit and hawq_rm_rejectrequest_nseg_limit to a percentage value

+++ src/include/cdb/cdbvars.h
@@ -1190,8 +1190,8 @@ extern int	   rm_segment_config_refresh_interval;
 extern int	   rm_segment_tmpdir_detect_interval;
 
 extern int	   rm_nocluster_timeout;
-extern int	   rm_tolerate_nseg_limit;
-extern int	   rm_rejectrequest_nseg_limit;
+extern double  rm_tolerate_nseg_limit;
+extern double  rm_rejectrequest_nseg_limit;
 extern int	   rm_nvseg_variance_among_seg_limit;
 extern int	   rm_container_batch_limit;
 extern char   *rm_resourcepool_test_filename;
